pyproject.toml
[tool.poetry]
name = "ai-add-to-story"
version = "0.1.0"
description = ""
authors = ["indivar <indimishra@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
crewai = "*"
langchain-google-genai = "*"
python-dotenv = "^1.0.1"
elevenlabs = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
